Read Write Atomic Access
Back to basic read write access. The read and write single operations are wrapped in the critical section so blocking doesn't occur for long if simultaneous access is attempted. The double buffer had the problems of the buffers getting out of sync so this option is being explored. It would be better to have operator overloads similar to the normal atomic.